ai-learning-dashboard/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
openai==1.47.0
tiktoken>=0.8.0
numpy>=2.1.0
scipy>=1.14.0
pydantic==2.9.2
python-multipart==0.0.12
httpx==0.27.2
